rename bitcoin class name
